blog create refactor

+++ helpers/imageUpload.js
@@ -26,25 +26,23 @@ const imageUploads = (files, next) => {
       Key: `news/${date}${files.originalname}`
     };
 
-    fs.unlinkSync(files.path);
 
-    // s3.upload(params, (err, data) => {
-    //   if (err) {
-    //     //   console.log('Error occured while trying to upload to S3 bucket', err);
-    //     return next({
-    //       err: err
-    //     });
-    //   }
+    s3.upload(params, (err, data) => {
+      if (err) {
+        //   console.log('Error occured while trying to upload to S3 bucket', err);
+        return next({
+          err: err
+        });
+      }
 
-    //   if (data) {
-    //     fs.unlinkSync(files.path); // Empty temp folder
-    //     const locationUrl = data.Location;
-    //     return next({
-    //       data: locationUrl
-    //     });
-    //   }
-    // });
-
+      if (data) {
+        fs.unlinkSync(files.path); // Empty temp folder
+        const locationUrl = data.Location;
+        return next({
+          data: locationUrl
+        });
+      }
+    });
   } catch (error) {
     return errs(error);
   }
